# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Datris API
  description: |
    REST API for the Datris AI Data Platform. Ingest, validate, transform, store, and retrieve data.

    For AI agent integration, use the [MCP Server](https://docs.datris.ai/mcp-server) instead.
  version: 1.28.0
  contact:
    name: Datris
    url: https://datris.ai
  license:
    name: Apache 2.0
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Optional API key for authentication (enabled via application.yaml)
  schemas:
...
    TapConfig:
      type: object
      required: [name, description, targetPipeline]
      properties:
        name:
          type: string
          description: Unique tap identifier
        description:
          type: string
          description: Plain-English instruction used to generate the script
        scriptPath:
          type: string
          description: MinIO path of the generated Python script (built-in storage)
        scriptStorage:
          type: string
          enum: [minio, github]
          description: Script storage backend. Absent or `minio` = built-in object storage; `github` = the tenant's configured code repository.
        scriptRepoPath:
          type: string
          description: Repo-relative script path when `scriptStorage` is `github`
        scriptCommitSha:
          type: string
          description: Commit SHA the script is pinned to when `scriptStorage` is `github`; runs read exactly this commit
        scriptKind:
          type: string
          enum: [python, http]
          description: Tap implementation kind. Absent or `python` = a script the platform executes; `http` = a user-hosted endpoint speaking the tap HTTP contract (`endpointUrl` is authoritative; script/packages fields do not apply).
        endpointUrl:
          type: string
          description: For `scriptKind` `http` only — absolute http(s) URL Datris POSTs the run context to on each run
        source:
          type: string
          description: Declared source identity for lineage and provenance — a provider name or host (never credentials). Optional; when absent the platform derives it from the endpoint host (HTTP taps) or the host the script references most. On update, omitting the field keeps the stored value and an empty string clears it.
        targetPipeline:
          type: string
          description: Pipeline this tap feeds into (may be empty for unattached taps)
        packages:
          type: array
          items:
            type: string
          description: Extra pip packages required by the script
        secretName:
          type: string
          description: Name of the tap secret (Vault) injected as env vars at runtime
        cronExpression:
          type: string
          description: >-
            Quartz CRON expression for scheduled runs (omit for manual-only). Six
            fields — seconds minutes hours day-of-month month day-of-week — plus an
            optional year, e.g. `0 0 * * * ?` (hourly). A 5-field Unix cron such as
            `*/1 * * * *` is rejected with 400; the error quotes the 6-field equivalent.
        enabled:
          type: boolean
          default: true
        tapType:
          type: string
          enum: [structured, document]
          default: structured
          description: >-
            `structured` (default) — the script returns records that flow into a
            structured pipeline destination.
            `document` — the script returns `{uri, filename, content}` dicts
            destined for a vector-store pipeline (unstructured source +
            qdrant/pgvector/weaviate/milvus/chroma destination). The server
            rejects a document tap pointed at an incompatible pipeline with
            HTTP 400. Document taps also use a per-URI ledger so each file is
            processed once across re-runs.
        lastRunStatus:
          type: string
          enum: [success, failure]
        lastRunTime:
          type: string
        lastRunRecordCount:
          type: integer
        lastRunError:
          type: string
        lastRunDataType:
          type: string
          enum: [csv, json, xml, text, document]
        lastRunColumns:
          type: array
          items:
            type: string
        lastTestRunStatus:
          type: string
        lastTestRunTime:
          type: string
        lastTestRunRecordCount:
          type: integer
        lastTestRunError:
          type: string
        lastTestRunDataType:
          type: string
        lastTestRunScriptId:
          type: string
          readOnly: true
          description: |
            Server-stamped identity of the exact script (or endpoint) that last passed a
            `mode=test` run (`gh:<sha>`, `http:<url>`, or sha256 of the stored script).
            Compared against the saved script when a `cronExpression` is set. Ignored in
            request bodies; null on pre-existing taps.
        lastTestRunColumns:
          type: array
          items:
            type: string
        createdAt:
          type: string
        updatedAt:
          type: string
...
    EntityVersionSummary:
      type: object
      properties:
        version:
          type: integer
        createdAt:
          type: string
        createdBy:
          type: string
          description: Who saved this version (user, API key label, or agent identity)
        changeNote:
          type: string
...
    EntityVersionDiff:
      type: object
      properties:
        from:
          type: integer
          description: Base version of the comparison (the `against` parameter)
        to:
          type: integer
          description: Target version of the comparison (the `version` parameter)
        config:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
              before:
                type: string
              after:
                type: string
              change:
                type: string
        script:
          type: object
          description: Script diff (taps only)
          properties:
            before:
              type: string
            after:
              type: string
            lines:
              type: array
              items:
                type: object
                properties:
                  type:
                    type: string
                  text:
                    type: string
paths:
...
  /api/v1/tap/versions:
    get:
      tags: [Versions]
      summary: List tap definition versions
      description: Lists the saved definition versions of a tap, newest first. Every versioned save (create, update, restore) appends a snapshot.
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
          description: Tap name
      responses:
        '200':
          description: Version summaries, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EntityVersionSummary'
        '500':
          description: Error
...
  /api/v1/tap/version/diff:
    get:
      tags: [Versions]
      summary: Diff two tap versions
      description: Server-computed, ready-to-render diff of the tap configuration and script between two versions.
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
        - name: version
          in: query
          required: true
          schema:
            type: integer
          description: Target version of the comparison
        - name: against
          in: query
          required: true
          schema:
            type: integer
          description: Base version to compare against
      responses:
        '200':
          description: Field-level config diff and line-level script diff
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntityVersionDiff'
        '404':
          description: One of the versions does not exist
        '500':
          description: Error
  /api/v1/tap/version/restore:
    post:
      tags: [Versions]
      summary: Restore a tap version
      description: >-
        Restores the tap's DEFINITION from the given version snapshot,
        preserving the live tap's creation metadata, ownership, and run-status
        fields. The restore itself is saved as a new version.
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
        - name: version
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: The restored tap configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TapConfig'
        '404':
          description: No such version for this tap
        '409':
          description: >-
            The snapshot's script differs from the one that last passed a test while a
            cron is set (test-before-cron). Body is `{"error": "<tap name + remedy>"}`.
        '500':
          description: Error
  /api/v1/pipeline/versions:
    get:
      tags: [Versions]
      summary: List pipeline definition versions
      description: Lists the saved definition versions of a pipeline, newest first.
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
          description: Pipeline name
      responses:
        '200':
          description: Version summaries, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EntityVersionSummary'
        '500':
          description: Error
...
  /api/v1/pipeline/version/diff:
    get:
      tags: [Versions]
      summary: Diff two pipeline versions
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
        - name: version
          in: query
          required: true
          schema:
            type: integer
          description: Target version of the comparison
        - name: against
          in: query
          required: true
          schema:
            type: integer
          description: Base version to compare against
      responses:
        '200':
          description: Field-level config diff (no script section)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntityVersionDiff'
        '404':
          description: One of the versions does not exist
        '500':
          description: Error
  /api/v1/pipeline/version/restore:
    post:
      tags: [Versions]
      summary: Restore a pipeline version
      description: Restores the pipeline's definition from the given version snapshot, preserving ownership. The restore itself is saved as a new version.
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
        - name: version
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: The restored pipeline configuration
          content:
            application/json:
              schema:
                type: object
        '404':
          description: No such version for this pipeline
        '500':
          description: Error

  # ── Authentication & Users ─────────────────────────────────
...